Further cleanup of code, and caught a memory leak each time the game is lost.

+++ trunk/MineView+Drawing.m
@@ -11,38 +11,36 @@
 @implementation MineView (Drawing)
 
 
-- (void)drawCellAtRow: (int) r column: (int) c inRect: (NSRect) rect {
-	Cell * cell = [field cellAtRow: r column: c];
+- (void)drawCellAtRow:(int)r column:(int)c inRect:(NSRect)rect {
+	Cell *cell = [field cellAtRow:r column:c];
     
 	if (state == gameLose) {
-		if ( (mousePoint.x == c) && (mousePoint.y == r) ) {
+		if ((mousePoint.x == c) && (mousePoint.y == r)) {
 			[[NSColor redColor] set];
-			[NSBezierPath fillRect: rect];
-			[self drawCellClearedBackgroundInRect: rect];
-			[self drawMineInRect: rect];
-		}
-		else if ( [cell isMine] && ![cell isFlagged] ) {
+			[NSBezierPath fillRect:rect];
+			[self drawCellClearedBackgroundInRect:rect];
+			[self drawMineInRect:rect];
+		} else if ([cell isMine] && ![cell isFlagged]) {
 			[[colors colorWithKey:@"background"] set];
-			[NSBezierPath fillRect: rect];
-			[self drawCellHiddenBackgroundInRect: rect];
-			[self drawMineInRect: rect];
-		}
-        else if ( ![cell isMine] && [cell isFlagged] ) {
+			[NSBezierPath fillRect:rect];
+			[self drawCellHiddenBackgroundInRect:rect];
+			[self drawMineInRect:rect];
+		} else if (![cell isMine] && [cell isFlagged]) {
             [[colors colorWithKey:@"background"] set];
-            [NSBezierPath fillRect: rect];
-            [self drawCellHiddenBackgroundInRect: rect];
-            [self drawFlagInRect: rect];
-            [self drawXInRect: rect];
+            [NSBezierPath fillRect:rect];
+            [self drawCellHiddenBackgroundInRect:rect];
+            [self drawFlagInRect:rect];
+            [self drawXInRect:rect];
         }
 		return;
 	}
 	
 	[[colors colorWithKey:@"background"] set];
-	[NSBezierPath fillRect: rect];
+	[NSBezierPath fillRect:rect];
     
 	if (drag == leftClick) {
-        int x = (int)mousePoint.x;
-        int y = (int)mousePoint.y;
+        int x = mousePoint.x;
+        int y = mousePoint.y;
 		if ( x==c && y==r && [cell isHidden]) {
 			[self drawCellClearedBackgroundInRect: rect];
 			return;
@@ -58,82 +56,80 @@
 		}
     }*/
     
-	NSString * text = nil;
-	NSColor * color = [NSColor blackColor];	
+	NSString *text = nil;
+	NSColor *color = [NSColor blackColor];	
 	
 	if ([cell isCleared]) {
-		[self drawCellClearedBackgroundInRect: rect];
+		[self drawCellClearedBackgroundInRect:rect];
 		int n = [cell neighbors];
-		text = [NSString stringWithFormat: @"%d", n];
-		color = [colors colorWithKey: text];
-	}
-	else {
-		[self drawCellHiddenBackgroundInRect: rect];
+		text = [NSString stringWithFormat:@"%d", n];
+		color = [colors colorWithKey:text];
+	} else {
+		[self drawCellHiddenBackgroundInRect:rect];
 		if ([cell isFlagged]) {
 			[self drawFlagInRect: rect];
-		}
-		else if ([cell isQuestioned]) {
+		} else if ([cell isQuestioned]) {
 			text = @"?";
 		}
 	}
     
 	if (text) {
 		/* Set up attributes dictionary */
-		NSFont * font = [NSFont fontWithName: @"Helvetica-Bold" size: 12]; 
-		NSDictionary * attrs = [NSDictionary dictionaryWithObjectsAndKeys:
+		NSFont *font = [NSFont fontWithName:@"Helvetica-Bold" size:12]; 
+		NSDictionary *attrs = [NSDictionary dictionaryWithObjectsAndKeys:
 			font, NSFontAttributeName,
 			color, NSForegroundColorAttributeName, nil];
 		
 		/* Center the drawing horizontally */
-		NSSize size = [text sizeWithAttributes: attrs];
+		NSSize size = [text sizeWithAttributes:attrs];
 		NSPoint location;
-		location.x = (rect.size.width-size.width)/2 + rect.origin.x + 1;
+		location.x = (rect.size.width - size.width)/2 + rect.origin.x + 1;
 		location.y = rect.origin.y;
 		
 		/* Draw the text */
-		[text drawAtPoint: location withAttributes: attrs];
+		[text drawAtPoint:location withAttributes:attrs];
 	}
     
 	[cell updated];
 }
 
-- (void)drawCellClearedBackgroundInRect: (NSRect) rect {
+- (void)drawCellClearedBackgroundInRect:(NSRect)rect {
     NSInteger n = (NSInteger)(cellHeight / 16.0f);
     
-	NSBezierPath * topLeft = [NSBezierPath bezierPath];
-	[topLeft moveToPoint: NSMakePoint(rect.origin.x,rect.origin.y)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x,rect.origin.y+rect.size.height)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+rect.size.width,rect.origin.y+rect.size.height)];	
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+rect.size.width-n,rect.origin.y+rect.size.height-n)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+n,rect.origin.y+rect.size.height-n)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+n,rect.origin.y+n)];
+	NSBezierPath *topLeft = [NSBezierPath bezierPath];
+	[topLeft moveToPoint:NSMakePoint(rect.origin.x, rect.origin.y)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x, rect.origin.y + rect.size.height)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + rect.size.width, rect.origin.y + rect.size.height)];	
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + rect.size.width - n, rect.origin.y + rect.size.height - n)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + n, rect.origin.y + rect.size.height - n)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + n, rect.origin.y + n)];
 	[topLeft closePath];
     
 	[[colors colorWithKey:@"shadow"] set];
 	[topLeft fill];
 }
 
-- (void)drawCellHiddenBackgroundInRect: (NSRect) rect {
+- (void)drawCellHiddenBackgroundInRect:(NSRect)rect {
 	NSInteger n = (NSInteger)(cellHeight / 8.0f);
 	
-	NSBezierPath * topLeft = [NSBezierPath bezierPath];
+	NSBezierPath *topLeft = [NSBezierPath bezierPath];
 	/* Build the top Left corner */ 
-	[topLeft moveToPoint: NSMakePoint(rect.origin.x,rect.origin.y)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x,rect.origin.y+rect.size.height)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+rect.size.width,rect.origin.y+rect.size.height)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+rect.size.width-n,rect.origin.y+rect.size.height-n)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+n,rect.origin.y+rect.size.height-n)];
-	[topLeft lineToPoint: NSMakePoint(rect.origin.x+n,rect.origin.y+n)];
+	[topLeft moveToPoint:NSMakePoint(rect.origin.x, rect.origin.y)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x, rect.origin.y + rect.size.height)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + rect.size.width, rect.origin.y + rect.size.height)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + rect.size.width - n, rect.origin.y + rect.size.height - n)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + n, rect.origin.y + rect.size.height - n)];
+	[topLeft lineToPoint:NSMakePoint(rect.origin.x + n, rect.origin.y + n)];
 	[topLeft closePath];
 	
-	NSBezierPath * bottomRight = [NSBezierPath bezierPath];
+	NSBezierPath *bottomRight = [NSBezierPath bezierPath];
 	/* Build the bottom right corner */
-	[bottomRight moveToPoint: NSMakePoint(rect.origin.x,rect.origin.y)];
-	[bottomRight lineToPoint: NSMakePoint(rect.origin.x+rect.size.width,rect.origin.y)];
-	[bottomRight lineToPoint: NSMakePoint(rect.origin.x+rect.size.width,rect.origin.y+rect.size.height)];
-	[bottomRight lineToPoint: NSMakePoint(rect.origin.x+rect.size.width-n,rect.origin.y+rect.size.height-n)];
-	[bottomRight lineToPoint: NSMakePoint(rect.origin.x+rect.size.width-n,rect.origin.y+n)];
-	[bottomRight lineToPoint: NSMakePoint(rect.origin.x+n,rect.origin.y+n)];
+	[bottomRight moveToPoint:NSMakePoint(rect.origin.x, rect.origin.y)];
+	[bottomRight lineToPoint:NSMakePoint(rect.origin.x + rect.size.width, rect.origin.y)];
+	[bottomRight lineToPoint:NSMakePoint(rect.origin.x + rect.size.width, rect.origin.y + rect.size.height)];
+	[bottomRight lineToPoint:NSMakePoint(rect.origin.x + rect.size.width - n, rect.origin.y + rect.size.height - n)];
+	[bottomRight lineToPoint:NSMakePoint(rect.origin.x + rect.size.width - n, rect.origin.y + n)];
+	[bottomRight lineToPoint:NSMakePoint(rect.origin.x + n, rect.origin.y + n)];
 	[bottomRight closePath];
 	
 	/* draw the two paths*/
@@ -144,15 +140,15 @@
 	[bottomRight fill];
 }
 
-- (void)drawFlagInRect: (NSRect) rect {
+- (void)drawFlagInRect:(NSRect)rect {
 	float nH = rect.size.height/16;
 	float nW = rect.size.width/16;
-	float base = rect.origin.y+(nH*3);
-	float top = rect.origin.y+(nH*13);
-	float flagHeight = nH*4;
-	float flagWidth = nW*5;
+	float base = rect.origin.y + (nH * 3);
+	float top = rect.origin.y + (nH * 13);
+	float flagHeight = nH * 4;
+	float flagWidth = nW * 5;
 	
-	NSBezierPath * flagPole = [NSBezierPath bezierPath];
+	NSBezierPath *flagPole = [NSBezierPath bezierPath];
 	/* Build the flag pole */
 	[flagPole moveToPoint: NSMakePoint(rect.origin.x+(nW*4),base)];
 	[flagPole lineToPoint: NSMakePoint(rect.origin.x+(nW*12),base)];
@@ -179,15 +175,15 @@
 
 - (void)drawMineInRect: (NSRect) rect {
 	NSRect inset = NSInsetRect(rect, cellWidth/4, cellHeight/4);
-	NSBezierPath * mine = [NSBezierPath bezierPathWithOvalInRect: inset];
-	NSBezierPath * hLine = [NSBezierPath bezierPathWithRect: NSInsetRect(rect, 2, 7.5f)];
-	NSBezierPath * vLine = [NSBezierPath bezierPathWithRect: NSInsetRect(rect, 7.5f, 2)];
+	NSBezierPath *mine = [NSBezierPath bezierPathWithOvalInRect:inset];
+	NSBezierPath *hLine = [NSBezierPath bezierPathWithRect:NSInsetRect(rect, 2, 7.5f)];
+	NSBezierPath *vLine = [NSBezierPath bezierPathWithRect:NSInsetRect(rect, 7.5f, 2)];
     
 	inset.size.height = cellHeight/8;
 	inset.size.width = cellWidth/8;
 	inset.origin.y += cellHeight/4;
 	inset.origin.x += cellHeight/16;
-	NSBezierPath * light = [NSBezierPath bezierPathWithOvalInRect: inset];
+	NSBezierPath *light = [NSBezierPath bezierPathWithOvalInRect:inset];
     
     [[NSColor blackColor] set];
     [mine fill];
@@ -199,20 +195,20 @@
 }
 
 - (void)drawXInRect: (NSRect) rect {
-    NSRect r = NSInsetRect(rect,cellWidth/8.0f,cellHeight/8.0f);
+    NSRect r = NSInsetRect(rect, cellWidth/8.0f, cellHeight/8.0f);
     
-    NSBezierPath * down = [NSBezierPath bezierPath];
-    [down moveToPoint: NSMakePoint(r.origin.x, r.origin.y+r.size.height*15.0f/16.0f)];
-    [down lineToPoint: NSMakePoint(r.origin.x+r.size.width/16.0f, r.origin.y+r.size.height)];
-    [down lineToPoint: NSMakePoint(r.origin.x+r.size.width, r.origin.y+r.size.height/16.0f)];
-    [down lineToPoint: NSMakePoint(r.origin.x+r.size.width*15.0f/16.0f, r.origin.y)];
+    NSBezierPath *down = [NSBezierPath bezierPath];
+    [down moveToPoint:NSMakePoint(r.origin.x, r.origin.y + r.size.height * 15.0f / 16.0f)];
+    [down lineToPoint:NSMakePoint(r.origin.x+r.size.width/16.0f, r.origin.y+r.size.height)];
+    [down lineToPoint:NSMakePoint(r.origin.x+r.size.width, r.origin.y+r.size.height/16.0f)];
+    [down lineToPoint:NSMakePoint(r.origin.x+r.size.width*15.0f/16.0f, r.origin.y)];
     [down closePath];
     
-    NSBezierPath * up = [NSBezierPath bezierPath];
-    [up moveToPoint: NSMakePoint(r.origin.x, r.origin.y+r.size.height/16.0f)];
-    [up lineToPoint: NSMakePoint(r.origin.x+r.size.width*15.0f/16.0f, r.origin.y+r.size.height)];
-    [up lineToPoint: NSMakePoint(r.origin.x+r.size.width, r.origin.y+r.size.height*15.0f/16.0f)];
-    [up lineToPoint: NSMakePoint(r.origin.x+r.size.width/16.0f, r.origin.y)];
+    NSBezierPath *up = [NSBezierPath bezierPath];
+    [up moveToPoint:NSMakePoint(r.origin.x, r.origin.y+r.size.height/16.0f)];
+    [up lineToPoint:NSMakePoint(r.origin.x+r.size.width*15.0f/16.0f, r.origin.y+r.size.height)];
+    [up lineToPoint:NSMakePoint(r.origin.x+r.size.width, r.origin.y+r.size.height*15.0f/16.0f)];
+    [up lineToPoint:NSMakePoint(r.origin.x+r.size.width/16.0f, r.origin.y)];
     [up closePath];
     
     [[NSColor redColor] set];
